+++ services/sso/src/middleware/authenticate.ts
@@ -1,9 +1,10 @@
 import { Context, Next } from 'koa'
 import jwt, { JwtPayload } from 'jsonwebtoken'
-import { InvalidCredentials } from '../utils/errors'
+import { ForbiddenError, InvalidCredentials } from '../utils/errors'
 import { appConfig } from '../config'
 import { client } from '../models/db'
 import { ValidateSchema } from '../schemas'
+import { UserStatus } from '../schemas/users/userSchema'
 
 export const authenticate = async (ctx: Context, next: Next) => {
   const { authToken } = ctx.request.body as ValidateSchema
@@ -12,12 +13,15 @@ export const authenticate = async (ctx: Context, next: Next) => {
   }
   const { id } = jwt.verify(authToken, appConfig.jwtKey) as JwtPayload
   const userResult = await client.query(
-    'SELECT id, role FROM "user" WHERE id = $1',
+    'SELECT id, role, status FROM "user" WHERE id = $1',
     [id]
   )
   const user = userResult.rows[0]
   if (!user) {
     throw new InvalidCredentials()
+  }
+  if (user.status === UserStatus.BLOCKED) {
+    throw new ForbiddenError('The user is Blocked')
   }
   ctx.state.user = user
   await next()
@@ -30,12 +34,15 @@ export const authenticateCookie = async (ctx: Context, next: Next) => {
   }
   const { id } = jwt.verify(authToken, appConfig.jwtKey) as JwtPayload
   const userResult = await client.query(
-    'SELECT id, role FROM "user" WHERE id = $1',
+    'SELECT id, role, status FROM "user" WHERE id = $1',
     [id]
   )
   const user = userResult.rows[0]
   if (!user) {
     throw new InvalidCredentials()
+  }
+  if (user.status === UserStatus.BLOCKED) {
+    throw new ForbiddenError('The user is Blocked')
   }
   ctx.state.user = user
   await next()