+++ web/wiki/src/__tests__/molecules/UserButton.test.tsx
@@ -1,7 +1,8 @@
 import { vi } from 'vitest'
-import { render, screen } from '../test-utils'
+import { render, screen, fireEvent, waitFor } from '../test-utils'
 import { TAuthContext, useAuth } from '../../context/AuthProvider'
 import { UserButton } from '../../components/molecules'
+import { paths } from '../../constants'
 
 vi.mock('../../context/AuthProvider', async () => {
   const actual = await vi.importActual('../../context/AuthProvider')
@@ -13,14 +14,34 @@ vi.mock('../../context/AuthProvider', async () => {
   }
 })
 
+const navigate = vi.fn()
+
+vi.mock('react-router-dom', async () => {
+  const actual: Record<number, unknown> = await vi.importActual(
+    'react-router-dom'
+  )
+  return {
+    ...actual,
+    useNavigate: vi.fn(() => navigate),
+  }
+})
+
 describe('UserButton', () => {
-  it('should render the component and handle dropdown menu', () => {
+  it('opens login modal when avatar is clicked and user is not authenticated', () => {
     vi.mocked(useAuth).mockReturnValue({
       user: null,
     } as TAuthContext)
+
+    render(<UserButton />)
+
+    const avatar = screen.getByAltText('Avatar')
+    expect(avatar).toBeInTheDocument()
+
+    fireEvent.click(avatar)
+    expect(screen.getByTestId('loginModal')).toBeInTheDocument()
   })
 
-  it('shows avatar image if user is set', () => {
+  it('navigates to profile when avatar is clicked and user is authenticated', async () => {
     vi.mocked(useAuth).mockReturnValue({
       user: {
         name: 'User',
@@ -29,6 +50,13 @@ describe('UserButton', () => {
     } as TAuthContext)
 
     render(<UserButton />)
-    expect(screen.getByTestId('avatarImageUser')).toBeInTheDocument()
+
+    const avatarAuthenticated = screen.getByAltText('Avatar authenticated')
+    expect(avatarAuthenticated).toBeInTheDocument()
+
+    fireEvent.click(avatarAuthenticated)
+    await waitFor(() => {
+      expect(navigate).toHaveBeenCalledWith(paths.profile)
+    })
   })
 })