+++ web/wiki/src/styles/index.ts
@@ -1,5 +1,0 @@
-export { colors } from './colors'
-export { dimensions } from './dimensions'
-export { FlexBox } from './FlexBox'
-export { font } from './font'
-export { responsiveSizes, device } from './mediaQueries'