+++ web/wiki/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.15.0] - 2024-03-12
+
+### Removed
+
+- Import missing atoms and styles from UI
+- Removed old atoms, styles and tests
+
+## [0.14.0] - 2024-03-11
+
+### Removed
+
+- Import molecules from UI
+- Removed old molecules and tests
+
 ## [0.13.1] - 2024-03-07
 
 ### Fixed